+++ src/components/withNuomi/index.jsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import invariant from 'invariant';
-import { getLocation } from '../../core/router';
 import { NuomiContext } from '../Context';
 
 const withNuomi = (WrapperComponent) => {
@@ -11,17 +10,16 @@ const withNuomi = (WrapperComponent) => {
 
     constructor(...args) {
       super(...args);
-      const { nuomiProps, nuomiRouteProps } = this.context;
+      const { nuomiProps } = this.context;
       invariant(
         nuomiProps,
         `不允许在 <Route>、<Nuomi>、<NuomiRoute> 外部使用 ${WithNuomi.displayName}`,
       );
-      this.location = nuomiRouteProps ? nuomiRouteProps.location : getLocation();
     }
 
     render() {
       const { nuomiProps } = this.context;
-      return <WrapperComponent {...this.props} nuomiProps={nuomiProps} location={this.location} />;
+      return <WrapperComponent {...this.props} nuomiProps={nuomiProps} />;
     }
   };
 };