+++ src/hooks/useNuomi.js
@@ -1,23 +1,11 @@
-import { useContext, useRef } from 'react';
+import invariant from 'invariant';
+import { useContext } from 'react';
 import { NuomiContext } from '../components/Context';
-import { getLocation } from '../core/router';
 
 const useNuomi = () => {
-  const { nuomiProps, nuomiRouteProps } = useContext(NuomiContext);
-  if (nuomiRouteProps) {
-    return nuomiRouteProps;
-  }
-
-  const cacheProps = useRef();
-  // 防止每次组件更新都重新计算
-  if (!cacheProps.current) {
-    cacheProps.current = {
-      ...nuomiProps,
-      location: getLocation(),
-    };
-  }
-
-  return cacheProps.current;
+  const { nuomiProps } = useContext(NuomiContext);
+  invariant(nuomiProps, '不允许在 <Route>、<Nuomi>、<NuomiRoute> 外部使用 useNuomi');
+  return { nuomiProps };
 };
 
 export default useNuomi;